+++ src/app.module.ts
@@ -8,8 +8,6 @@ import { FileManagerModule } from './module/file_manager.module';
 import configuration from './config/configuration';
 import dbConfig from './config/db.config';
 
-import { AuthenticationMiddleware } from './middleware/authentication.middleware';
-import * as Controllers from './controller';
 import { UserService } from './service/user.service';
 import { UserRepository } from './repository/user.repository';
 import { DatabaseConfig } from './config/interfaces';
@@ -29,17 +27,4 @@ import { DatabaseConfig } from './config/interfaces';
   providers: [UserService, UserRepository],
 })
 export class AppModule {
-  configure(consumer: MiddlewareConsumer) {
-    const controllers: Array<Type<any>> = Object.values(Controllers);
-
-    // TODO: list endpoint which requires auth
-    consumer
-      .apply(AuthenticationMiddleware)
-      .exclude(
-        { path: '/system/healthy', method: RequestMethod.GET },
-        { path: '/users/register', method: RequestMethod.POST },
-        { path: '/files/download/(.*)', method: RequestMethod.GET }
-      )
-      .forRoutes(...controllers);
-  }
 }
