+++ src/main.ts
@@ -3,21 +3,20 @@ import { FastifyAdapter, NestFastifyApplication } from '@nestjs/platform-fastify
 import { SwaggerModule, DocumentBuilder, SwaggerDocumentOptions } from '@nestjs/swagger';
 import { ConfigService } from '@nestjs/config';
 import { Logger } from '@nestjs/common';
-import { CorsOptions } from '@nestjs/common/interfaces/external/cors-options.interface';
 import fmp from 'fastify-multipart';
 
 import { AppModule } from './app.module';
 import { SwaggerConfig } from './config/interfaces';
 import { ErrorInterceptor } from './interceptor/error.interceptor';
+import { AuthenticationInterceptor } from './interceptor/authentication.interceptor';
+import { UserService } from './service/user.service';
 
 async function bootstrap() {
   const app = await NestFactory.create<NestFastifyApplication>(AppModule, new FastifyAdapter({ maxParamLength: 1000 }));
   const configService = app.get(ConfigService);
   const logger = new Logger(bootstrap.name);
 
-  const cors: CorsOptions = configService.get('swagger') as CorsOptions;
-
-  app.enableCors(cors);
+  app.enableCors();
   app.register(fmp);
 
   const swagger: SwaggerConfig = configService.get('swagger') as SwaggerConfig;
@@ -34,7 +33,8 @@ async function bootstrap() {
   const document = SwaggerModule.createDocument(app, config, options);
   SwaggerModule.setup('api', app, document);
 
-  app.useGlobalInterceptors(new ErrorInterceptor());
+  const userService = app.get(UserService);
+  app.useGlobalInterceptors(new AuthenticationInterceptor(userService), new ErrorInterceptor());
 
   const port = configService.get('port');
