+++ src/shared/request.ts
@@ -1,10 +1,6 @@
 import { User } from 'src/model';
-import { FastifyRequest, RawRequestDefaultExpression } from 'fastify';
-
-interface RawRequest extends RawRequestDefaultExpression {
-  user: User;
-}
+import { FastifyRequest } from 'fastify';
 
 export interface Request extends FastifyRequest {
-  raw: RawRequest;
+  user: User
 }
